+++ fetchql.config.example.js
@@ -2,7 +2,8 @@
 module.exports = {
     server: {
         host: 'SERVER_HOSTNAME',
-        token: 'API_BEARER_TOKEN'
+        token: 'API_BEARER_TOKEN',
+        perPage: 100
     },
     paths: {
         input: './PATH_TO_QUERIES_FILES',
@@ -11,6 +12,8 @@ module.exports = {
     env: {
         iterableValue: ['Value1', 'Value2', '...'],
         fixedValue: 'Value'
-    }
+    },
+    plugins: [],
+    queries: {}
 }